Require Go 1.22; enforce GET on /api/config/channels
It's not a good look that this—and other Go backends I've written—have
been this lax on HTTP methods.

There's a lot I could say here about my own attention to detail and
choice of tools, or perhaps about the entire philosophy underpinning the
Go language and ecosystem. But I'm going to avoid saying any of it,
because what's important is that this is fixed now, and that others have
already said some of those parts better than I ever will.

What I _will_ say is that when I first started the v2 rewrite of
Hypcast, I couldn't manage to get the stability I needed out of
GStreamer's WebRTC implementation, and the Rust port of Pion hadn't
seen its v0.1.0 release.