Simplify a bunch of client code
I'm honestly shocked at how much complexity I put into the header
implementation, with layers upon layers of functions obscuring the logic
behind the various status indicators. Was this one of those cases where
I did the whole "make all your functions as tiny as possible!" thing,
before I permanently concluded that it's a terrible idea? Or was this
more complex to start and I never refactored it as much as I needed to?