Don't limit color fades based on prefers-reduced-motion
I'm not sure what made me think this counted as "motion." I guess it was
more like: "here's an accessibility feature I just learned about; I must
apply it somehow, because accessibility works best when you don't think
too hard about the needs that drive it." Which is a bad way to (not)
think.

The docs for this on MDN give an example of a box with a growing and
shrinking animation changing to a box that dissolves in and out (i.e.
fades its color over time). Scaling and panning are where issues come
in, and I can't think of where such motion would even make sense in the
Hypcast UI.