Change github.com/gorilla/websocket version
This is a great example of why I've stopped aggressively bumping
dependencies for no good reason, and why it was morally wrong (†) for me
to do this across my projects in the first place. Apparently v1.5.1
introduced some print statements in library code that I was lucky enough
not to notice. I only found out after looking into the new v1.5.3
release, the tag for which implies it's a revert back to v1.5.0, but
which still has some diffs compared to v1.5.0 that I have yet to fully
review.

(† I'm being a bit sarcastic here.)