Note caveats about the RPC framework
I don't want to give the impression that this is the best example to
build on. Hypcast has a whole thing in the README about not exposing it
to the Internet, and how it's a personal project, so it should be well
understood that design decisions here and elsewhere reflect that.