Restore and revamp the watch benchmarks
I took this out when I added type parameterization to watch.Value,
calling it "ridiculous and useless" because "the results probably just
reflect lock contention."

Well, I'm about to change the locking strategy for watch.Value in a big
way, so a head-to-head comparison of Value.Set implementations in the
presence of watchers has some meaning, even if it's all about lock
contention.

My original implementation really was awful, though. Every setter
incrementing the same atomic integer? Really?

This version introduces three big changes from the original:

1. Adding a version with 1,000 watchers.
2. Using type parameters.
3. Using random numbers (from the math/rand/v2 default source) to
   generate the values to set.

The random part is still fuzzy to me. It's not like I've proven that
using a constant is worse than this, so the idea that the setters should
do "something" is basically just vibes. But the goal is to compare
implementations, not look at absolute numbers for their own sake, so the
work done by setters isn't *that* important either way.