Introduce Tprintf to simplify the current log format
Hypcast previously had a bunch of individual "logf" methods to dump out
a log prefixed by a type name and pointer address. This change uses a
bit of reflection and a bit of rework on the existing type names to
extract that into one shared function.

There's an interesting mix of generics and reflection here. I'm using
the type parameter to enforce that src is a pointer, which guarantees
that "%p" will work right and simplifies getting the type name compared
to digging through layers of Elem() that will panic with a wrong move.

I'm considering moving Hypcast to log/slog. Even if I throw this away
soon, it's nice that I'll be able to search for "log." across the whole
codebase to see what needs to change.