Introduce the embedclient build tag and strongly deprecate embedclientzip
Explained in the comment. I've actually known for a while that HTTP file
serving requires the files to implement io.Seeker, but still hadn't
bothered to fix this.

This transitional commit introduces the new* build tag alongside the
"zip" one, to permit head-to-head comparisons and I guess test a gradual
update of the build process. However, I intend to remove the old build
tag in the very next commit.

(* It's actually the original build tag from before I moved to the
ZIP-based approach.)