Improve TypeScript configuration
On another project that uses Vite, I found a number of settings that
seemed out of alignment with Vite's real behavior. For example, Vite
targets ES2020 by default, but TypeScript was allowing ESNext on the
client.

All of this *should* be more consistent and future-proof, though I have
to admit I'm still wrapping my head around the realities of modules in
the world of JavaScript.