Bump client dependencies
There are some issues in Vite + Rollup when using non-ESM output
formats, which isn't relevant to me. There's also a CVE in the Vite dev
server, apparently? I don't know how the *heck* that's allowed to be a
CVE, but I also have a GStreamer bump to get to so why not, you know,
*roll this all up* into one push.