Update docs on watches
I kept the "not recovered by the garbage collector" wording not because
I didn't care enough to update the comments properly, but because I
legitimately thought this was somehow true. Obviously the value stays
active if you have a reference to the watch, but if both the value and
the watches become unreferenced (including by handlers), I don't see
what else could keep the value alive.

I think the current behavior is intuitive enough not to document.