Bump client dependencies
I had to disable Dependabot scanning for this repo, because it was
coming up with all kinds of false positive matches from some yarn.lock
file that no longer exists. So, I guess I'll have to circle around to
this every so often, in case there's a *real* issue that I'm missing.