Bump eslint-plugin-react-hooks to v5.0.0
There's no default flat config, but this version lets you wire things up
manually without using the compatibility shims. See the issue in the
React GitHub repository whose ID is the product of 23 and 1231, which I
don't want GitHub to auto-link via a full web URL because the issue is
crowded enough as-is.

(Also, because I have a specific history of being an awful person and
linking an angrily written commit message from this very project into a
different Facebook-originated OSS project while working around a bug
that wasn't even that bad. Since this was in Hypcast v1, the relevant
change and my note about rewriting the repo history are no longer
accessible from this repo's current branches. However, my behavior in
this incident is shameful and atrocious enough that a record of its
existence deserves to remain here.)