Use the semi-official GitHub mirror for GStreamer
See comments in the change for details.

I can't find any reference to this repo from an official GStreamer
source, but I've always understood the "Public mirror" badge on GitHub
to be special, and not something anyone can put on their repos.

Besides my desire to shift the bandwidth and CPU burden of this shallow
Git clone from (presumably) a team of volunteers onto one of the world's
largest corporations, I figure it's faster and cheaper for a GitHub
Actions runner to clone from here than from an arbitrary place on the
Internet.