Create testing/synctest versions of some watch tests
The "testing/synctest" package, available behind a GOEXPERIMENT in
Go 1.24, is a vastly more robust alternative to the things I currently
use my forceRuntimeProgress() helper for.

This initial set of tests covers everything that currently uses that
helper, whether on its own or through assertBlocked(). The changes
compared to the original tests are minimal; I'm sure I'll experiment
further over time.

Since the new package is behind a GOEXPERIMENT (and only in a Go release
candidate as of writing), I have of course hidden these tests behind the
build tag for that experiment. They will not replace any existing tests
until the synctest API is finalized and made available in a stable Go
release.