Eliminate custom assertions from synctest examples
Most of the point of these custom assertions is to end the test more
quickly than the default 10 minute timeout in the event of a deadlock.
With testing/synctest, deadlocks panic and crash the test immediately.

This change also calls out that several of the custom assertions use
testing.T.Fatal, which is perfectly valid in a normal test but not
technically valid under a synctest bubble (which is rooted in a
goroutine distinct from that of the test function).